# openapi/arena-api.yaml
openapi: 3.1.0
info:
  title: LoL Arena Stats API
  version: 0.1.0
  description: Contract-first API for a Rust backend that serves arena stats.
servers:
  - url: http://localhost:3001
paths:
  /health:
    get:
      operationId: getHealth
      tags:
        - system
      summary: Backend health check
      responses:
        '200':
          description: Service health
          content:
            application/json:
              schema:
                $ref: ../schemas/health-status.json
  /api/v1/arena/stats:
    get:
      operationId: getArenaStats
      tags:
        - arena
      summary: Fetch aggregated arena stats for a player
      parameters:
        - name: playerName
          in: query
          required: true
          schema:
            type: string
        - name: tagLine
          in: query
          required: false
          schema:
            type: string
        - name: region
          in: query
          required: false
          schema:
            type: string
            default: americas
      responses:
        '200':
          description: Aggregated arena stats snapshot
          content:
            application/json:
              schema:
                $ref: ../schemas/arena-player-stats.json
  /lol/players/search:
    get:
      operationId: searchPlayers
      tags:
        - players
      summary: Search players by Riot ID game name and tagline
      parameters:
        - name: game_name
          in: query
          required: true
          schema:
            type: string
        - name: tagline
          in: query
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: ../schemas/player-search-response.json
        '502':
          description: Upstream Riot/API failure
          content:
            application/json:
              schema:
                $ref: ../schemas/error-response.json
  /lol/players/{puuid}:
    get:
      operationId: getPlayerProfile
      tags:
        - players
      summary: Get player profile and recent Arena matches
      parameters:
        - name: puuid
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Player profile with recent matches
          content:
            application/json:
              schema:
                $ref: ../schemas/player-profile-response.json
        '404':
          description: Player not found
          content:
            application/json:
              schema:
                $ref: ../schemas/error-response.json
        '502':
          description: Upstream Riot/API failure
          content:
            application/json:
              schema:
                $ref: ../schemas/error-response.json